Move AbstractionSolverInstance to cegar folder + fix includes
